Renamed module file name to reflect that this is for generic use, not use for graph rag
