+++ src/backend/base/langflow/components/vectorstores/astradb.py
@@ -5,8 +5,8 @@ from astrapy import AstraDBAdmin, DataAPIClient, Database
 from astrapy.info import CollectionDescriptor
 from langchain_astradb import AstraDBVectorStore, CollectionVectorServiceOptions
 
-from langflow.base.vectorstores.graph_rag_decorator import vector_store_connection
 from langflow.base.vectorstores.model import LCVectorStoreComponent, check_cached_vector_store
+from langflow.base.vectorstores.vector_store_connection_decorator import vector_store_connection
 from langflow.helpers import docs_to_data
 from langflow.inputs import FloatInput, NestedDictInput
 from langflow.io import (
